# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gym Project API
  version: 1.0.0
  description: JSON REST API interacting with the gym project backend
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-production-url.com/api
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: auth_key
      description: Authentication key obtained from /auth/login
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
...
    UserProfile:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        username:
          type: string
        phone:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        password:
          type: string
paths:
...
  /users/profile:
    get:
      summary: Get authenticated user profile
      tags:
        - Users
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/me:
    patch:
      summary: Update current user's profile
      tags:
        - Users
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpdateProfileRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ==================== SESSIONS ROUTES ====================
...
security:
  - ApiKeyAuth: []